requirements.txt
Django
djangorestframework
pylance